backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1

sqlalchemy==2.0.25
psycopg2-binary==2.9.9

pydantic==2.6.1
python-multipart==0.0.9
python-dotenv==1.0.1

spacy==3.7.4
scikit-learn==1.4.1.post1
pandas==2.2.0
numpy==1.26.4

PyMuPDF==1.23.25
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz

# Enhanced NLP & ML
torch==2.1.0
transformers==4.36.0
sentence-transformers==2.2.2
shap==0.44.0
scikit-learn==1.4.1.post1
xgboost==2.0.0
lightgbm==4.1.0

# Vector Database
pinecone-client==2.2.4
weaviate-client==3.26.1

# Async Processing
celery==5.3.4
redis==5.0.1

# Advanced PDF Processing
pdfplumber==0.10.3
camelot-py==0.11.0  # For table extraction

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0